# [Wait for #3457] Android-Optimized Q4_0 OpenCL Kernel