--- a/jni/Android.mk.in
+++ b/jni/Android.mk.in
@@ -260,8 +260,12 @@ LOCAL_MODULE_TAGS   := optional
 LOCAL_LDLIBS        := -llog -landroid -fopenmp -static-openmp
 LOCAL_LDFLAGS 	    += "-Wl,-z,max-page-size=16384"
 
-@MESON_OPENCL_SHARED_LIBS@
-LOCAL_STATIC_LIBRARIES += iniparser openblas ruy @MESON_OPENCL_STATIC_LIBS@
+LOCAL_STATIC_LIBRARIES += iniparser openblas ruy
+
+ifeq (@MESON_ENABLE_OPENCL@,1)
+  LOCAL_SHARED_LIBRARIES := OpenCL
+  LOCAL_STATIC_LIBRARIES += clblast
+endif
 
 ifeq ($(MESON_HAS_TFLITE), 1)
   LOCAL_STATIC_LIBRARIES += tensorflow-lite