… nntrainer/tensor/cpu_backend/fallback/fallback_internal.h
… test/jni/Android.mk
… test/jni/meson.build
… test/unittest/unittest_blas_kernels_cl.cpp
… tools/android_test.sh