This PR introduces an Android-optimized Q4_0 OpenCL kernel that enables OpenCL image acceleration on Android devices.
The function for unpacking q4_0x4 data to use on the GPU is newly implemented and accelerated using NEON.

**Self-evaluation:**
1. Build test: [X]Passed [ ]Failed [ ]Skipped
2. Run test:   [X]Passed [ ]Failed [ ]Skipped


### Result

#### 1. 3072 x 8192 x 3072
|engine|avg|
|------|---|
|CPU|266.6 ms|
|GPU|165.8 ms|

#### 2. 3072 x 3072 x 8192 
|engine|avg|
|------|---|
|CPU|248 ms|
|GPU|150.2 ms|

#### 3. 3072 x 3072 x 3072 
|engine|avg|
|------|---|
|CPU|109.4 ms|
|GPU|67.2 ms|

### Note

This benchmark was tested on the Galaxy S23 Ultra.