@dkjung 2025-09-09
Does GPU kernel for Android use Q40x4, not Q40x8?