@djeong20 2025-09-09
> Does GPU kernel for Android use Q4_0x4, not Q4_0x8?

CPU uses Q4_0x4 on Android. Therefore, a different unpack function is needed to support this.